Added option 'count' and implemented web.busy
This patch adds the ability to specify the amount of challenges the user
wants to recieve. This is done using the 'count' argument. The count
cannot be higher than 10.

The busy variable:
* The bot sets the busy variable True whenever it is doing something
* If a command is sent while the bot is busy, it gets replied with a
message stating that the bot is busy
* The busy variable is a requirement for the count argument to work
properly and also improves multi-user functionality
* Without the busy flag the bot could break whenever it recieved two
commands siultaneously and the likelyhood of that is greater when the
users have the ability to use the 'count' argument