Added support for battle-royale-distance and powerups
This patch adds support for the (new) battle-royale-distance game mode
It works by specifying "mode=brd" together with the link to the lobby.

Battle-royale rules can now be specified using the "rules=" option.
With the rules the user can specify whether any powerups are enabled or
not.

Please note that none of the powerups exist in the
battle-royale-distance gamemode.